+++ changelog.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Moved `fetchCSS` and `fetchJS` into one file
+-   Moved `fetchCSS` and `fetchJS` into one file [#10](https://github.com/Pageworks/djinnjs/issues/10)
 -   Both fetch functions create their own loading animation tickets
 
 ### Fixed